# swagger-spec.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        "200":
          description: ""
  /auth/signup:
    post:
      operationId: AuthController_signup
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpDto"
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseEntityDto"
      tags:
        - auth
  /auth/signin:
    post:
      operationId: AuthController_signin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: ""
      tags:
        - auth
  /auth/verify:
    post:
      operationId: AuthController_verify
      parameters:
        - name: code
          required: true
          in: query
          schema:
            type: string
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - auth
  /auth/change-password:
    post:
      operationId: AuthController_changePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordDto"
      responses:
        "201":
          description: ""
      tags:
        - auth
  /auth/change-username:
    post:
      operationId: AuthController_changeUsername
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeUsernameDto"
      responses:
        "201":
          description: ""
      tags:
        - auth
  /auth/username-available:
    post:
      operationId: AuthController_usernameAvailable
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckUsernameDto"
      responses:
        "201":
          description: ""
      tags:
        - auth
  /auth/email-available:
    post:
      operationId: AuthController_emailAvailable
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckEmailDto"
      responses:
        "201":
          description: ""
      tags:
        - auth
  /users/ban:
    post:
      operationId: UsersController_banUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserDto"
      responses:
        "200":
          description: ""
      tags:
        - users
  /users/unban:
    post:
      operationId: UsersController_unbanUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnlyUserIdDto"
      responses:
        "200":
          description: ""
      tags:
        - users
  /users:
    get:
      operationId: UsersController_getAllUsers
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - users
  "/users/{userId}":
    get:
      operationId: UsersController_getUserById
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - users
    delete:
      operationId: UsersController_deleteUser
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - users
  /roles/change:
    post:
      operationId: RolesController_changeUserRole
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRoleDto"
      responses:
        "200":
          description: ""
      tags:
        - roles
  /roles:
    get:
      operationId: RolesController_getAvailableRoles
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - roles
info:
  title: Dogtor API
  description: Documentation for Dogtor API
  version: "1.0"
  contact: {}
tags: []
servers: []
components:
  schemas:
    SignUpDto:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - username
        - email
        - password
    ResponseEntityDto:
      type: object
      properties:
        statusCode:
          type: number
        message:
          type: string
        token:
          type: string
        data:
          type: object
      required:
        - statusCode
        - message
        - token
        - data
    LoginDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    ChangePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      required:
        - oldPassword
        - newPassword
    ChangeUsernameDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    CheckUsernameDto:
      type: object
      properties:
        username:
          type: string
      required:
        - username
    CheckEmailDto:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    BanUserDto:
      type: object
      properties: {}
    OnlyUserIdDto:
      type: object
      properties: {}
    ChangeRoleDto:
      type: object
      properties: {}